Manually merging Sherif's refactoring
