+++ apacheasyncbenchmark/src/test/java/com/ss/benchmark/httpclient/apacheasync/ApacheAsyncBasePerformanceTest.java
@@ -1,18 +1,18 @@
-package com.ss.benchmark.httpclient.reactornetty;
+package com.ss.benchmark.httpclient.apacheasync;
 
 import com.ss.benchmark.httpclient.common.HttpClientEngine;
-import com.ss.benchmark.httpclient.common.PerformanceTests;
+import com.ss.benchmark.httpclient.common.BasePerformanceTest;
 import org.testng.annotations.Test;
 
 /**
  * @author sharath.srinivasa
  */
 @Test(groups = "performance")
-public class ReactorNettyPerformanceTests extends PerformanceTests {
+public class ApacheAsyncBasePerformanceTest extends BasePerformanceTest {
 
     @Override
     protected HttpClientEngine getClient() {
-        return new ReactorNettyEngine();
+        return new ApacheAsyncEngine();
     }
 }
