+++ asynchttpclient-benchmark/src/test/java/benchmark/AsyncHttpClientBasePerformanceTest.java
@@ -1,18 +1,18 @@
 package benchmark;
 
 import com.ss.benchmark.httpclient.common.HttpClientEngine;
-import com.ss.benchmark.httpclient.common.PerformanceTests;
+import com.ss.benchmark.httpclient.common.BasePerformanceTest;
 import org.testng.annotations.Test;
 
 /**
  * @author sharath.srinivasa
  */
 @Test(groups = "performance")
-public class RXNettyPerformanceTests extends PerformanceTests {
+public class AsyncHttpClientBasePerformanceTest extends BasePerformanceTest {
 
     @Override
     protected HttpClientEngine getClient() {
-        return new RXNettyEngine();
+        return new AsyncHttpClientEngine();
     }
 }
