+++ reactornettybenchmark/src/test/java/com/ss/benchmark/httpclient/reactornetty/ReactorNettyBasePerformanceTest.java
@@ -1,18 +1,18 @@
-package benchmark;
+package com.ss.benchmark.httpclient.reactornetty;
 
 import com.ss.benchmark.httpclient.common.HttpClientEngine;
-import com.ss.benchmark.httpclient.common.PerformanceTests;
+import com.ss.benchmark.httpclient.common.BasePerformanceTest;
 import org.testng.annotations.Test;
 
 /**
  * @author sharath.srinivasa
  */
 @Test(groups = "performance")
-public class AsyncHttpClientPerformanceTests extends PerformanceTests {
+public class ReactorNettyBasePerformanceTest extends BasePerformanceTest {
 
     @Override
     protected HttpClientEngine getClient() {
-        return new AsyncHttpClientEngine();
+        return new ReactorNettyEngine();
     }
 }
